backend/requirements.txt
pip
Django
asgiref
sqlparse
psycopg2
psycopg2-binary
djangorestframework
setuptools
django-cors-headers
djangorestframework-simplejwt
stripe
google-auth
requests
python-decouple
boto3
